backend/requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.2
django-extensions>=3.2
ipython
flake8
